% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioneaCaso.R
\name{FunzioneaCaso}
\alias{FunzioneaCaso}
\title{funzione a caso}
\usage{
FunzioneaCaso(x)
}
\arguments{
\item{x}{number to be cubed.}
}
\description{
This function is extremely useless.
}
\keyword{noreferences;}
