% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTriangle.R
\name{print.Triangle}
\alias{print.Triangle}
\title{Basic print function}
\usage{
\method{print}{Triangle}(obj)
}
\arguments{
\item{obj}{object we take stuff from.}
}
\description{
This function allows you to print the triangle area.
}
\keyword{geometry;}
\keyword{triangle;}
