% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangle.R
\name{Triangle}
\alias{Triangle}
\title{A basic triangle}
\usage{
Triangle(base = NULL, altezza = NULL, triD = TRUE)
}
\arguments{
\item{base}{Triangle base Default set to NULL}

\item{altezza}{Triangle height Default set to NULL}
}
\description{
This function allows you to express your love of cats.
}
\examples{
piastrella<-Triangle(base=2,altezza = 5)
print(piastrella)

}
\references{
Plat, R. (2009). On stochastic mortality modeling. Insurance: Mathematics 
  and Economics, 45(3), 393-404. 

Hunt, A., & Blake, D. (2015). On the Structure and Classification of 
Mortality Models Mortality Models. Pension Institute Working Paper.
\url{http://www.pensions-institute.org/workingpapers/wp1506.pdf}.
}
\keyword{geometry;}
\keyword{triangle;}
